WIP: Use GitHub workflows for deployment
Travis CI has been decommed for a while.
We just haven't needed a go.uber.org deployment for that period.

Switch to GitHub workflows to manage these deployments.

Ceremony involved if it ever becomes relevant again:

- Set up a service account "gh-pages" in Google Cloud using the instructions at https://cloud.google.com/iam/docs/creating-managing-service-accounts

- Grant this account the following roles per https://github.com/google-github-actions/deploy-appengine#via-google-github-actionsauth:
  roles/appengine.appAdmin, roles/iam.serviceAccountUser, roles/compute.storageAdmin, roles/cloudbuild.builds.editor, roles/cloudscheduler.admin

- Grant the following additional roles:

    - roles/storage.objectCreator to address

        ```
        "gh-actions@go-uber-org.iam.gserviceaccount.com does not have storage.objects.create access to the Google Cloud Storage object"
        ```

- Export a long-lived account key JSON using instructions at https://cloud.google.com/iam/docs/creating-managing-service-account-keys.

- Paste the contents of this JSON into a GitHub secret at https://github.com/uber/go.uber.org/settings/secrets/actions.

Following that, the workflow YAML is pretty straightforward.

While this is WIP,
I've made the following two changes to the deployment:

- Build on push to this branch
- Do *not* promote to serve all traffic

When this lands on master, I'll be changing those.

Note that because promotion to serve production traffic is optional,
this theoretically lets us have PR previews in the future that do not
get promoted.